requirements.txt
python-telegram-bot==20.7
yt-dlp==2024.3.10
python-dotenv==1.0.0
aiofiles==23.2.1
apscheduler==3.10.4